mark-track/backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic[email]
python-dotenv